Extract host reports tests into integration_core_hosts_reports_test.go
Relates to #51839

Move 16 host-report/vitals test methods (TestGetMacadminsData,
TestHostsReportDownload, TestGetHostDiskEncryption, TestOSVersions,
TestHostHealth, etc.) out of integration_core_test.go into their own
file.
